refactor(dates): reorganise the dates pipeline
